# gitrank/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GitRank Gateway API
  version: 0.2.0
  description: |
    Public and authenticated HTTP surface served by `api-gateway`.

    Compatibility policy:
    - all externally supported routes are versioned under `/v1`
    - additive changes may add optional fields without a version bump
    - breaking changes require a new versioned route family and a documented migration window

    Conventions:
    - authenticated browser clients use the `gitrank_session` cookie
    - state-changing routes additionally require `X-CSRF-Token`
    - public profile responses are cacheable for a short window
    - authenticated profile and sync responses are returned with `Cache-Control: private, no-store`
    - rate limiting currently exposes `Retry-After` on `429 Too Many Requests`

    Pagination and filtering:
    - current `/v1` routes do not expose paginated collections
    - future collection routes should use `page[size]`, `page[after]`, `page[before]`, `sort`, and `filter[...]`
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/sync:
    post:
      tags: [Sync]
      summary: Request a GitHub sync job
      description: |
        Accepts a sync request and normalizes it to one of the implemented GitHub ingestor sync routes.

        Idempotency:
        - this endpoint does not currently accept an explicit idempotency key
        - repeated requests may enqueue additional work if they fall outside downstream dedupe rules
        - webhook-driven syncs are deduplicated separately by GitHub delivery ID
      operationId: createSyncRequest
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
            examples:
              syncUser:
                summary: Sync the authenticated GitHub user
                value:
                  mode: user
              syncRepository:
                summary: Sync a repository
                value:
                  mode: repository
                  repository: openai/gpt-oss
              syncPullRequest:
                summary: Sync a pull request
                value:
                  mode: pull_request
                  repository: gitrank/gitrank
                  number: 42
      responses:
        "202":
          description: Sync job accepted
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "400":
          description: Invalid sync request or JSON payload
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid or missing CSRF token
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream auth or ingestion dependency failed
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/sync/repository/execute:
    post:
      tags: [Sync]
      summary: Execute a bounded live repository sync
      description: |
        Fetches recent public repository, pull request, review, issue, and commit data directly through the GitHub ingestor
        and persists it through the normalized PostgreSQL model.

        This route is intentionally bounded and is not a historical backfill API.
      operationId: executeRepositorySync
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
            example:
              mode: repository
              repository: gitrank/gitrank
      responses:
        "200":
          description: Repository sync completed
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitHubSyncExecutionResponse"
        "400":
          description: Invalid repository sync request or JSON payload
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid or missing CSRF token
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: GitHub ingestor could not complete the live repository sync
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/sync/installation/execute:
    post:
      tags: [Sync]
      summary: Execute a bounded installation sync from persisted installation associations
      description: |
        Replays repositories already associated with a persisted GitHub installation record and delegates each one to the
        bounded repository executor.

        This route does not perform live GitHub App installation discovery and is intentionally scoped to the OAuth-only
        v1 production baseline.
      operationId: executeInstallationSync
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
            example:
              mode: installation
              installation_id: 42
      responses:
        "200":
          description: Installation sync completed
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitHubSyncExecutionResponse"
        "400":
          description: Invalid installation sync request or JSON payload
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid or missing CSRF token
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: GitHub ingestor could not complete the bounded installation sync
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/me/account/unlink:
    post:
      tags: [Account]
      summary: Disconnect the linked GitHub account and clear the current session
      operationId: unlinkMyAccount
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        "200":
          description: Linked account disconnected and current session cleared
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountUnlinkResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid or missing CSRF token
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/me/account/export:
    get:
      tags: [Account]
      summary: Export authenticated account-owned GitRank data without token secrets
      description: |
        Returns the authenticated user's profile, score history, badges, repository visibility,
        linked GitHub account metadata, sanitized session metadata, and sanitized audit events.
        Raw session tokens, token hashes, CSRF hashes, encrypted GitHub tokens, and refresh tokens
        are intentionally excluded.
      operationId: exportMyAccountData
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Account data export generated
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDataExportResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/me/account/delete:
    post:
      tags: [Account]
      summary: Delete the authenticated GitRank account and clear the current session
      description: |
        This route performs confirmed self-service account deletion for the authenticated user.

        Current v1 behavior:
        - user-owned GitRank records are hard-deleted immediately after confirmation
        - detached repository-level contribution evidence may remain when historical foreign keys use `ON DELETE SET NULL`
      operationId: deleteMyAccount
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
            examples:
              confirmDelete:
                summary: Confirm account deletion
                value:
                  confirmation: DELETE
      responses:
        "200":
          description: Account deleted and current session cleared
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeletionResponse"
        "400":
          description: Invalid confirmation or JSON payload
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid or missing CSRF token
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/me/profile:
    get:
      tags: [Profiles]
      summary: Return the authenticated user's profile read model
      operationId: getMyProfile
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current authenticated profile snapshot
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateProfileResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Profiles]
      summary: Update authenticated profile privacy controls
      operationId: updateMyProfilePrivacy
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfilePrivacyRequest"
      responses:
        "200":
          description: Updated authenticated profile snapshot
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateProfileResponse"
        "400":
          description: Invalid privacy update
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid or missing CSRF token
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/me/quests:
    get:
      tags: [Profiles]
      summary: Return authenticated quest recommendations derived from profile score evidence
      description: |
        Quest recommendations are materialized from the latest profile snapshot and verified score history.
        Completed quests write progress, completion, reward grant, badge, score-event, and audit evidence.
      operationId: getMyQuests
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current authenticated quest board
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserQuestsResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/leaderboard:
    get:
      tags: [Profiles]
      summary: Return the public GitRank leaderboard
      description: |
        The leaderboard is built from public, leaderboard-enabled profile snapshots materialized into
        weekly leaderboard season snapshots and rank movement events.
        Placement is directional and formula-versioned, not an absolute measure of engineering ability.
      operationId: getLeaderboard
      responses:
        "200":
          description: Public leaderboard snapshot
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PublicProfileCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  headers:
    XRequestID:
      description: Request correlation ID injected at the gateway edge.
      schema:
        type: string
...
    PrivateNoStoreCacheControl:
      description: Private no-store cache policy applied to authenticated and sync responses.
      schema:
        type: string
        example: private, no-store
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: gitrank_session
      description: "Session cookie issued by `auth-service`."
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit CSRF token derived from the current session.
  schemas:
...
    SyncResponse:
      type: object
      required: [status, accepted_at]
      properties:
        status:
          type: string
          example: queued
        job_id:
          type: string
        correlation_id:
          type: string
        accepted_at:
          type: string
          format: date-time
...